Fix progress not being applied properly onto the loading bar
The noop `$loading` interface accounts for `start` but it's never used. As a result, no matter the duration of the request and its progress, it'd stay invisible and only appear at the very end of the request with a 100% progress.

The behavior can be observed in this [reproduction link](https://codesandbox.io/s/2ol8m8j9mr). Inside about.vue'd `fetch`, progress works just fine. When you click on Request 1 though, the loading "blinks". When you click on Request 2, which does call start, it works.

Thank you both @bovas85 and @pimlie for helping me out.