feat: create axios instance on process.axios for server-middleware use
An extra instance of axios is created and attached to "process.axios" so
that server middlewares have access to the axios instance that has
extensions provided by this module.

The "proxyHeaders" option is ignored for that instance as it would be
both awkward to implement and potentially wrong. This is because
"proxyHeaders" is proxying requests from the request that initiated the
page request while server-middleware instance is kinda independent of
individual page requests.

Also, it's not an entirely perfect solution because that "process.axios"
instance is only extended when axios server-side plugin is executed.
Ideally, it would be extended immediately when the module's code runs.
That would require some refactoring to eliminate templates from the
code that extends the instance and move it to a separate file. This is
doable but would result in fewer byte savings.

Resolves #427